Add Free Mode button functionality to home screen
