+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,7 +9,7 @@ EditorBuildSettings:
     path: Assets/Scenes/HOME.unity
     guid: f062934542c517e4c96f75deb17a6c7f
   - enabled: 1
-    path: Assets/Scenes/GameScene.unity
+    path: Assets/Scenes/FREEMODE.unity
     guid: 99c9720ab356a0642a771bea13969a05
   m_configObjects:
     com.unity.input.settings.actions: {fileID: -944628639613478452, guid: 052faaac586de48259a63d0c4782560b, type: 3}
